Refactor: Eliminate code duplication and inconsistent dice rolling logic
## Changes

### Added
- `src/shared/search-content.ts`: Shared module for search/fetch content
  - Extracted searchable content and content library to single source of truth
  - Added `searchContent()` and `fetchContent()` helper functions

### Modified
- `src/index.ts`: Refactored to use shared search content module
  - Removed ~400 lines of duplicated search/fetch content
  - Now imports `searchContent` and `fetchContent` from shared module

- `app/[transport]/route.ts`: Complete rewrite to use core dice logic
  - Removed ~670 lines of simplified inline dice rolling implementation
  - Now uses `DiceNotationParser` and `DiceRoller` from core modules
  - Ensures consistent behavior between local (STDIO) and remote (HTTP) servers
  - Maintains all functionality with proper validation and error handling

## Benefits

1. **Single Source of Truth**: Search content maintained in one place
2. **Consistent Behavior**: Both local and remote servers use same parser/roller
3. **Better Validation**: Remote server now has full validation from core parser
4. **Reduced Maintenance**: Changes to dice logic only need to be made once
5. **Code Reduction**: Eliminated 1,070 lines of duplicate code

## Testing

- All 66 tests pass
- Linting passes with no errors
- TypeScript compilation successful
- No breaking changes to public API
